Refactored draggables into GuiHelpers.
