+++ CMakeLists.txt
@@ -390,7 +390,6 @@ set_property(TARGET Engine PROPERTY CXX_STANDARD_REQUIRED ON)
 set(EDITOR_SRCS
         src/Editor/main.cpp
         src/Editor/Editor.cpp
-        src/Editor/ImGui/Draggable.cpp
         src/Editor/ImGui/OpenGLImplementation.cpp
         src/Editor/ImGui/Splitter.cpp
         src/Editor/ImGui/Theme.cpp
@@ -411,7 +410,6 @@ set(EDITOR_SRCS
 
 set(EDITOR_HEADERS
         src/Editor/Editor.hpp
-        src/Editor/ImGui/Draggable.hpp
         src/Editor/ImGui/OpenGLImplementation.hpp
         src/Editor/ImGui/Splitter.hpp
         src/Editor/ImGui/Theme.hpp
